Fixed more unit tests
